Experimental changes for gear truck tilt angle
We use the gear compression for the MLG to ascertain the tilt of the truck. This will work regardless of the orientation of the aircraft in pitch and roll, and also on uneven surfaces.

Also added is a second set of main gear slightly in front of the current set and slightly higher - this is to simulate the z forces from the additional set of wheels touching. Again this might need tuning.

The numbers that probably need tuning.

* libraries.xml Gear Tilt 1 and Gear Tilt 2 interpolation table; value of 0.9 could be adjusted
* A330-300-GE.xml - X & Z positions of the Gear.